Apply fixes
